add specialized abstract parser for agents
